feat(serialization): support, and reccomend, 64 bit wide ngen::serialization_size in models
A 32 bit compatible path is proposed in this commit, allowing models to use
a narrower value for the serialized byte size up to INT32_MAX. Overflow becomes
a model concern in this case.

The outlier here is the Fortran ABI and the iso C interface, which *can*
support a 64 bit c int declaration on the model side, but requires marshalling
the size across the set_value_int function as an array of 4 byte ints, and using
`transfer` to correction assemble them.

Test coverage for each model expands to round trip into both a 32 bit and 64 bit
integer in the test models.

+++ src/utilities/bmi/serialization.cpp
@@ -141,7 +141,7 @@ auto NgenSerializationProtocol::run(const ModelPtr& model, const Context& ctx) c
         return {};
     }
 
-    int size = 0;
+    int64_t size = 0;
     try {
         // (create, free) scope is RAII-managed so any throw from
         // GetValue, the record build, or the backend write still
